requirements.txt
aws-cdk.core
aws-cdk.aws-ec2